update head light

+++ header/head.h
@@ -4,9 +4,9 @@
 #include "primitives.h"
 
 void drawRobotHead(double x, Prop3D props);
-void drawRobotEye(float x, Prop3D props);
+void drawRobotEye(double x, Prop3D props);
 void drawRobotCheek(double x, Prop3D props); 
-void drawRobotMouth(float x, Prop3D props);
+void drawRobotMouth(double x, Prop3D props);
 void drawRobotForehead(double x, Prop3D props);
 void drawRobotLeftEar(double x, Prop3D props);
 void drawRobotRightEar(double x, Prop3D props);
